Move the StorageConfig
