+++ pipelex/system/configuration/configs.py
@@ -1,7 +1,7 @@
 from typing import cast
 
 import shortuuid
-from pydantic import Field, field_validator, model_validator
+from pydantic import Field, field_validator
 
 from pipelex.base_exceptions import PipelexConfigError
 from pipelex.cogt.config_cogt import Cogt
@@ -14,7 +14,8 @@ from pipelex.system.configuration.config_model import ConfigModel
 from pipelex.system.configuration.config_root import ConfigRoot
 from pipelex.tools.aws.aws_config import AwsConfig
 from pipelex.tools.log.log_config import LogConfig
-from pipelex.types import Self, StrEnum
+from pipelex.tools.storage.storage_config import StorageConfig
+from pipelex.types import StrEnum
 
 
 class ConfigPaths:
@@ -33,11 +34,6 @@ class ConfigPaths:
     BASE_DECK_FILE_PATH = f"{MODEL_DECKS_DIR_PATH}/{BASE_DECK_FILE_NAME}"
     OVERRIDES_DECK_FILE_NAME = "overrides.toml"
     OVERRIDES_DECK_FILE_PATH = f"{MODEL_DECKS_DIR_PATH}/{OVERRIDES_DECK_FILE_NAME}"
-
-
-class StorageMethod(StrEnum):
-    LOCAL = "local"
-    IN_MEMORY = "in_memory"
 
 
 class ValidationErrorReaction(StrEnum):
@@ -61,56 +57,6 @@ class ValidationErrorConfig(ConfigModel):
                 value_enum_cls=ValidationErrorReaction,
             ),
         )
-
-
-class StorageLocalConfig(ConfigModel):
-    uri_format: str
-    local_storage_path: str
-
-
-class StorageInMemoryConfig(ConfigModel):
-    uri_format: str
-
-
-class StorageConfig(ConfigModel):
-    is_fetch_remote_content_enabled: bool
-    method: StorageMethod = Field(strict=False)
-    local: StorageLocalConfig | None = None
-    in_memory: StorageInMemoryConfig | None = None
-
-    @model_validator(mode="after")
-    def validate_storage_config(self) -> Self:
-        match self.method:
-            case StorageMethod.LOCAL:
-                if not self.local:
-                    msg = "local config is required when method is local"
-                    raise PipelexConfigError(msg)
-            case StorageMethod.IN_MEMORY:
-                if not self.in_memory:
-                    msg = "in_memory config is required when method is in_memory"
-                    raise PipelexConfigError(msg)
-        return self
-
-    @property
-    def storage_path(self) -> str:
-        if not self.local:
-            msg = "local config is required when method is local"
-            raise PipelexConfigError(msg)
-        return self.local.local_storage_path
-
-    @property
-    def uri_format(self) -> str:
-        match self.method:
-            case StorageMethod.LOCAL:
-                if not self.local:
-                    msg = "local config is required to access uri_format"
-                    raise PipelexConfigError(msg)
-                return self.local.uri_format
-            case StorageMethod.IN_MEMORY:
-                if not self.in_memory:
-                    msg = "in_memory config is required to access uri_format"
-                    raise PipelexConfigError(msg)
-                return self.in_memory.uri_format
 
 
 class PipeRunConfig(ConfigModel):
